opt: 元数据使用系统接口来获取，文件所处目录判断优化，文件更新监听处优化

+++ src/utils/fileWatcher.ts
@@ -16,12 +16,10 @@ export class FileWatcher {
   private metadataCache: MetadataCache;
   private callbacks: Set<FileChangeCallback> = new Set();
   private fileCache: Map<string, number> = new Map();
-  private dir: string;
 
   constructor(plugin: BanyanPlugin) {
     this.plugin = plugin;
     this.vault = plugin.app.vault;
-    this.dir = plugin.settings.cardsDirectory;
     this.metadataCache = plugin.app.metadataCache;
     this.handleCreate = this.handleCreate.bind(this);
     this.handleDelete = this.handleDelete.bind(this);
@@ -46,7 +44,7 @@ export class FileWatcher {
   }
 
   private async handleCreate(file: TFile) {
-    if (!this.plugin.fileUtils.isLegalMarkdownFile(file)) return;
+    if (!this || !this.plugin.fileUtils.isLegalMarkdownFile(file)) return;
     this.fileCache.set(file.path, file.stat.mtime);
     await ensureFileID(file, this.plugin.app);
     const fileInfo = createFileInfo(file, this.plugin.app);
@@ -55,13 +53,13 @@ export class FileWatcher {
   }
 
   private handleDelete(file: TFile) {
-    if (!this.plugin.fileUtils.isLegalMarkdownFile(file)) return;
+    if (!this || !this.plugin.fileUtils.isLegalMarkdownFile(file)) return;
     this.fileCache.delete(file.path);
     this.emit({ type: 'delete', fileInfo: file });
   }
 
   private handleModify(file: TFile) {
-    if (!this.plugin.fileUtils.isLegalMarkdownFile(file)) return;
+    if (!this || !this.plugin.fileUtils.isLegalMarkdownFile(file)) return;
     const fileInfo = createFileInfo(file, this.plugin.app);
     if (!fileInfo) return;
     this.fileCache.set(file.path, file.stat.mtime);
@@ -69,7 +67,7 @@ export class FileWatcher {
   }
 
   private handleRename(file: TFile, oldPath: string) {
-    if (!this.plugin.fileUtils.isLegalMarkdownFile(file)) return;
+    if (!this || !this.plugin.fileUtils.isLegalMarkdownFile(file)) return;
     const fileInfo = createFileInfo(file, this.plugin.app);
     if (!fileInfo) return;
     this.fileCache.set(file.path, this.fileCache.get(oldPath)!);
@@ -78,7 +76,7 @@ export class FileWatcher {
   }
 
   private handleMetaChange(file: TFile) {
-    if (!this.plugin.fileUtils.isLegalMarkdownFile(file)) return;
+    if (!this || !this.plugin.fileUtils.isLegalMarkdownFile(file)) return;
     const fileInfo = createFileInfo(file, this.plugin.app);
     if (!fileInfo) return;
     this.emit({ type: 'meta-change', fileInfo });
